Switched num_inputs in assertInputsAvailable from int to int64_t

+++ onnx/version_converter/helper.h
@@ -17,5 +17,5 @@ namespace ONNX_NAMESPACE { namespace version_conversion {
 
     void assertNotParams(const std::vector<Dimension>& sizes);
 
-    void assertInputsAvailable(const ArrayRef<Value*>& inputs, const char* name, int num_inputs);
+    void assertInputsAvailable(const ArrayRef<Value*>& inputs, const char* name, int64_t num_inputs);
 }}
